Controlador Update Fixes #161

+++ app/Http/Controllers/IncidenciaController.php
@@ -22,14 +22,22 @@ use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Storage;
 use Maatwebsite\Excel\Facades\Excel;
 
+use function PHPUnit\Framework\isNull;
+
 class IncidenciaController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index(Request $request)
+    public function index()
     {
-        $incidencias = Incidencia::all();
+        $user = auth()->user();
+
+        // Comprobamos si el rol del usuario es administrador.
+        // Si lo es, devolvemos todas las incidencias.
+        // En caso contrario (rol de profesor), devolvemos las incidencias creadas por el usuario.
+        $incidencias = $user->hasRole('administrador') ? Incidencia::all() : Incidencia::where('creador_id', $user->id)->get();
+
         return view('incidencias.index', compact('incidencias'));
     }
 
@@ -38,7 +46,15 @@ class IncidenciaController extends Controller
     */
     public function datosIncidencias()
     {
-        $incidenciasJSON = Incidencia::with(['subtipo', 'creador', 'equipo', 'comentarios'])->get();
+
+        $user = auth()->user();
+
+        // Comprobamos si el rol del usuario es administrador.
+        // Si lo es, devolvemos todas las incidencias.
+        // En caso contrario (rol de profesor), devolvemos las incidencias creadas por el usuario.
+        $incidenciasJSON = $user->hasRole('administrador') ? Incidencia::with(['subtipo', 'creador', 'equipo', 'comentarios'])->get() : Incidencia::where('creador_id', $user->id)->with(['subtipo', 'creador', 'equipo', 'comentarios'])->get();
+
+        //Devolvemos el JSON
         return response()->json($incidenciasJSON);
     }
 
@@ -59,10 +75,39 @@ class IncidenciaController extends Controller
     public function store(CrearIncidenciaRequest $request)
     {
         try {
-            DB::beginTransaction();
+            //Si el usuario no tiene email, le asignaremos el email qe inserte.
+            if (is_null(auth()->user()->email)) {
 
-            //Falta que si el usuario no tiene departamento asignado, al crear la incidencia se le muestre el departamento, y se le asigne el que elija
-            //Falta que si el usuario no tiene asignado correo , al crear una incidencia se le muestre el correo, y se le asigne el que introduzca
+                //Sacamos usuario en una variable
+                $user = auth()->user();
+
+                //Hacemos que el mail sea el introducido
+                $user->email = $request->email;
+
+                //Guardamos el user
+                $user->save;
+            }
+
+            //Si el usuario no tiene departamento_id o nombre_departamento, le asignaremos el departamento que elija.
+            if (is_null(auth()->user()->id_departamento) || is_null(auth()->user()->nombre_departamento)) {
+
+                //Sacamos usuario en una variable
+                $user = auth()->user();
+
+                //Buscamos el departamento por id y sacamos el objeto
+                $objDept = Departamento::where('id', $request->departamento)->get();
+
+                //Hacemos que el campo departamento id sea el introducido
+                $user->departamento_id = $request->departamento;
+
+                //Hacemos que el campo nombre de departamento sea el introducido
+                $user->nombre_departamento = $objDept->nombre_departamento;
+
+                //Guardamos el user
+                $user->save;
+            }
+
+            DB::beginTransaction();
 
             $incidencia = new Incidencia();
 
@@ -155,12 +200,10 @@ class IncidenciaController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(ModificarIncidenciaRequest $request, string $id)
+    public function update(ModificarIncidenciaRequest $request, Incidencia $incidencia)
     {
         try {
             DB::beginTransaction();
-
-            $incidencia = new Incidencia();
 
             $incidencia->tipo = $request->tipo;
 
@@ -186,15 +229,22 @@ class IncidenciaController extends Controller
             //$incidencia->fecha_creacion = now();
             $incidencia->descripcion = $request->descripcion;
             $incidencia->actuaciones = $request->actuaciones;
-            $incidencia->estado =  $request->estado;
+            //$incidencia->estado =  $request->estado;
+            $incidencia->estado =  "ABIERTA";//Hay que poner la de arriba
             $incidencia->prioridad = $request->prioridad;
 
             $incidencia->duracion = $request->duracion; //Solo sale cuando la incidencia esta resuelta o cerrada
 
+            //Comprobamos que el request traiga el campo fichero
             if ($request->hasFile('fichero')) {
-                Storage::disk('discoAssets')->delete($incidencia->adjunto_url);
+                //Eliminamos el fichero antiguo si existe
+                if (!is_null($incidencia->adjunto_url)) {
+                    Storage::disk('discoAssets')->delete($incidencia->adjunto_url);
+                }
+                // Almacenamos el nuevo fichero
                 $incidencia->adjunto_url = $request->file('fichero')->store('adjunto', 'discoAssets');
             } else {
+                // Si no hay un nuevo archivo, establecer el adjunto_url a null
                 $incidencia->adjunto_url = null;
             }
 
@@ -221,20 +271,42 @@ class IncidenciaController extends Controller
             // Envío de correo poniéndolo en cola para que no interrumpa la redirección
             //Mail::to([$incidencia->creador->nombre_completo])->queue(new EnvioCorreo($incidencia, 'creado'));
 
-            return redirect()->route('incidencias.show', compact('incidencia'))->with('success', 'Incidencia creada correctamente.');
+            return redirect()->route('incidencias.show', compact('incidencia'))->with('success', 'Incidencia modificada correctamente.');
         } catch (Exception $e) {
             dd($e);
             DB::rollBack();
-            return redirect()->route('incidencias.show', compact('incidencia'))->with('error', 'No se pudo crear la incidencia. Detalles: ' . $e->getMessage());
+            return redirect()->route('incidencias.show', compact('incidencia'))->with('error', 'No se pudo modificar la incidencia. Detalles: ' . $e->getMessage());
         }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(int $id)
     {
-        //
+        try {
+
+            // Buscar la incidencia por su ID
+            $incidencia = Incidencia::findOrFail($id);
+
+            // Iniciar la transacción
+            DB::beginTransaction();
+
+            // Eliminar la incidencia
+            $incidencia->delete();
+
+            // Confirmar la transacción
+            DB::commit();
+        } catch (Exception $e) {
+            // Deshacer la transacción en caso de error
+            DB::rollBack();
+
+            //Redirigiar al index con mensaje de error
+            return redirect()->route('incidencias.index')->with('error', 'Error al eliminar la incidencia. Detalles: ' . $e->getMessage());
+        }
+
+        // Redirigir al index después de la eliminación con mensaje de exito
+        return redirect()->route('incidencias.index')->with('success', 'Incidencia eliminada correctamente.');
     }
 
     /**
@@ -248,14 +320,14 @@ class IncidenciaController extends Controller
     {
         $fechaYHoraExportacion = date('YmdHis');
 
-        $claseExport = match($tipo) {
+        $claseExport = match ($tipo) {
             'todas' => IncidenciasExport::class,
             'resueltas' => IncidenciasResueltasAdministradoresExport::class,
             'abiertas' => IncidenciasAbiertasUsuarioExport::class,
             default => null,
         };
 
-        return match($formato) {
+        return match ($formato) {
             'pdf' => Excel::download(new $claseExport, $fechaYHoraExportacion . "_Incidencias_{$tipo}.pdf", \Maatwebsite\Excel\Excel::DOMPDF),
             'xlsx' => Excel::download(new $claseExport, $fechaYHoraExportacion . "_Incidencias_{$tipo}.xlsx"),
             'csv' => Excel::download(new $claseExport, $fechaYHoraExportacion . "_Incidencias_{$tipo}.csv", \Maatwebsite\Excel\Excel::CSV)
